refactor: former compute api/proxy full decommissioning
