+++ src/utils/index.ts
@@ -1,6 +1,5 @@
 export * as convert from './convert'
 export * as math from './math'
-export * as process from './process'
 export * as data from './data'
 export * as misc from './misc'
 export * as spatial from './spatial'
